+++ Domain/ForeignKeyTransformerInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Ivoz\Core\Application;
+namespace Ivoz\Core\Domain;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Ivoz\Core\Domain\Model\EntityInterface;
