+++ Domain/Helper/ArrayObjectHelper.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Ivoz\Core\Application\Helper;
+namespace Ivoz\Core\Domain\Helper;
 
 class ArrayObjectHelper
 {
