+++ Domain/Model/Changelog/ChangelogAbstract.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Ivoz\Core\Domain\Model\Changelog;
 
 use Assert\Assertion;
-use Ivoz\Core\Application\DataTransferObjectInterface;
+use Ivoz\Core\Domain\DataTransferObjectInterface;
 use Ivoz\Core\Domain\Model\ChangelogTrait;
 use Ivoz\Core\Domain\Model\EntityInterface;
-use Ivoz\Core\Application\ForeignKeyTransformerInterface;
+use Ivoz\Core\Domain\ForeignKeyTransformerInterface;
 use Ivoz\Core\Domain\Model\Helper\DateTimeHelper;
 use Ivoz\Core\Domain\Model\Commandlog\CommandlogInterface;
 use Ivoz\Core\Domain\Model\Commandlog\Commandlog;
