+++ Domain/Model/Changelog/ChangelogDtoAbstract.php
@@ -2,8 +2,8 @@
 
 namespace Ivoz\Core\Domain\Model\Changelog;
 
-use Ivoz\Core\Application\DataTransferObjectInterface;
-use Ivoz\Core\Application\Model\DtoNormalizer;
+use Ivoz\Core\Domain\DataTransferObjectInterface;
+use Ivoz\Core\Domain\Model\DtoNormalizer;
 use Ivoz\Core\Domain\Model\Commandlog\CommandlogDto;
 
 /**
