+++ Domain/Model/Changelog/ChangelogTrait.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Ivoz\Core\Domain\Model\Changelog;
 
-use Ivoz\Core\Application\DataTransferObjectInterface;
-use Ivoz\Core\Application\ForeignKeyTransformerInterface;
+use Ivoz\Core\Domain\DataTransferObjectInterface;
+use Ivoz\Core\Domain\ForeignKeyTransformerInterface;
 
 /**
 * @codeCoverageIgnore
