+++ Domain/Model/EntityInterface.php
@@ -2,8 +2,8 @@
 
 namespace Ivoz\Core\Domain\Model;
 
-use Ivoz\Core\Application\DataTransferObjectInterface;
-use Ivoz\Core\Application\ForeignKeyTransformerInterface;
+use Ivoz\Core\Domain\DataTransferObjectInterface;
+use Ivoz\Core\Domain\ForeignKeyTransformerInterface;
 
 /**
  * Entity interface
