+++ Domain/Service/Assembler/CustomEntityAssemblerInterface.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Ivoz\Core\Application\Service\Assembler;
+namespace Ivoz\Core\Domain\Service\Assembler;
 
-use Ivoz\Core\Application\DataTransferObjectInterface;
-use Ivoz\Core\Application\ForeignKeyTransformerInterface;
+use Ivoz\Core\Domain\DataTransferObjectInterface;
+use Ivoz\Core\Domain\ForeignKeyTransformerInterface;
 use Ivoz\Core\Domain\Model\EntityInterface;
 
 interface CustomEntityAssemblerInterface
