+++ Domain/Service/EntityPersisterInterface.php
@@ -2,7 +2,7 @@
 
 namespace Ivoz\Core\Domain\Service;
 
-use Ivoz\Core\Application\DataTransferObjectInterface;
+use Ivoz\Core\Domain\DataTransferObjectInterface;
 use Ivoz\Core\Domain\Model\EntityInterface;
 
 interface EntityPersisterInterface
