+++ Infrastructure/Application/DoctrineForeignKeyTransformer.php
@@ -4,21 +4,16 @@ namespace Ivoz\Core\Infrastructure\Application;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\EntityManagerInterface;
-use Ivoz\Core\Application\DataTransferObjectInterface;
-use Ivoz\Core\Application\ForeignKeyTransformerInterface;
-use Ivoz\Core\Application\Helper\EntityClassHelper;
+use Ivoz\Core\Domain\DataTransferObjectInterface;
+use Ivoz\Core\Domain\ForeignKeyTransformerInterface;
+use Ivoz\Core\Domain\Helper\EntityClassHelper;
 use Ivoz\Core\Domain\Model\EntityInterface;
 
 class DoctrineForeignKeyTransformer implements ForeignKeyTransformerInterface
 {
-    /**
-     * @var EntityManagerInterface
-     */
-    private $em;
-
-    public function __construct(EntityManagerInterface $em)
-    {
-        $this->em = $em;
+    public function __construct(
+        private EntityManagerInterface $em
+    ) {
     }
 
     /**
