+++ Infrastructure/Domain/Service/DoctrineEntityPersister.php
@@ -5,10 +5,10 @@ namespace Ivoz\Core\Infrastructure\Domain\Service;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\UnitOfWork;
-use Ivoz\Core\Application\DataTransferObjectInterface;
-use Ivoz\Core\Application\Helper\EntityClassHelper;
-use Ivoz\Core\Application\Service\CreateEntityFromDto;
-use Ivoz\Core\Application\Service\UpdateEntityFromDto;
+use Ivoz\Core\Domain\DataTransferObjectInterface;
+use Ivoz\Core\Domain\Helper\EntityClassHelper;
+use Ivoz\Core\Domain\Service\CreateEntityFromDto;
+use Ivoz\Core\Domain\Service\UpdateEntityFromDto;
 use Ivoz\Core\Domain\Model\EntityInterface;
 use Ivoz\Core\Domain\Service\EntityPersisterInterface;
 use Ivoz\Core\Infrastructure\Persistence\Doctrine\Events as CustomEvents;
