+++ Infrastructure/Domain/Service/Lifecycle/CommandPersister.php
@@ -2,8 +2,8 @@
 
 namespace Ivoz\Core\Infrastructure\Domain\Service\Lifecycle;
 
-use Ivoz\Core\Application\Event\CommandWasExecuted;
-use Ivoz\Core\Application\Service\CommandEventSubscriber;
+use Ivoz\Core\Domain\Event\CommandWasExecuted;
+use Ivoz\Core\Domain\Service\CommandEventSubscriber;
 use Ivoz\Core\Domain\Event\EntityEventInterface;
 use Ivoz\Core\Domain\Service\EntityEventSubscriber;
 use Ivoz\Core\Domain\Service\EntityPersisterInterface;
