+++ Infrastructure/Persistence/Redis/Lock.php
@@ -2,7 +2,7 @@
 
 namespace Ivoz\Core\Infrastructure\Persistence\Redis;
 
-use Ivoz\Core\Application\MutexInterface;
+use Ivoz\Core\Domain\MutexInterface;
 use Psr\Log\LoggerInterface;
 
 class Lock implements MutexInterface
