+++ Infrastructure/Symfony/DependencyInjection/Compiler/DomainServiceCompiler.php
@@ -2,7 +2,7 @@
 
 namespace Ivoz\Core\Infrastructure\Symfony\DependencyInjection\Compiler;
 
-use Ivoz\Core\Application\Helper\LifecycleServiceHelper;
+use Ivoz\Core\Domain\Helper\LifecycleServiceHelper;
 use Ivoz\Core\Domain\Service\DomainEventSubscriberInterface;
 use Ivoz\Core\Domain\Service\LifecycleEventHandlerInterface;
 use Ivoz\Core\Domain\Service\LifecycleServiceCollectionInterface;
